+++ docusign-sdk-sample-swift/docusign-sdk-sample-swift/Controllers/TemplatesViewController.swift
@@ -51,8 +51,7 @@ class TemplatesViewController: UIViewController, UITableViewDelegate, UITableVie
             }
         }
     }
-
-
+    
     // MARK: TableView Methods
     
     // number of sections in table view